Changed actions structure to remove code duplication
