+++ eisbw/src/main/java/eisbw/actions/Load.java
@@ -2,27 +2,13 @@ package eisbw.actions;
 
 import eis.iilang.Action;
 import eis.iilang.Numeral;
-import eis.iilang.Parameter;
 import jnibwapi.JNIBWAPI;
 import jnibwapi.Unit;
 
-import java.util.LinkedList;
-
-public class Load extends StarcraftAction {
+public class Load extends StarcraftLoadingAction {
 
   public Load(JNIBWAPI api) {
     super(api);
-  }
-
-  @Override
-  public boolean isValid(Action action) {
-    LinkedList<Parameter> parameters = action.getParameters();
-    return parameters.size() == 1 && parameters.get(0) instanceof Numeral;
-  }
-
-  @Override
-  public boolean canExecute(Unit unit, Action action) {
-    return unit.getType().getSpaceProvided() > 0;
   }
 
   @Override
