+++ eisbw/src/main/java/eisbw/actions/Patrol.java
@@ -9,22 +9,10 @@ import jnibwapi.Unit;
 
 import java.util.LinkedList;
 
-public class Patrol extends StarcraftAction {
+public class Patrol extends StarcraftMovableAction {
 
   public Patrol(JNIBWAPI api) {
     super(api);
-  }
-
-  @Override
-  public boolean isValid(Action action) {
-    LinkedList<Parameter> parameters = action.getParameters();
-    return parameters.size() == 2 && parameters.get(0) 
-        instanceof Numeral && parameters.get(1) instanceof Numeral;
-  }
-
-  @Override
-  public boolean canExecute(Unit unit, Action action) {
-    return unit.getType().isCanMove();
   }
 
   @Override
