+++ eisbw/src/main/java/eisbw/actions/UnloadUnit.java
@@ -2,27 +2,13 @@ package eisbw.actions;
 
 import eis.iilang.Action;
 import eis.iilang.Numeral;
-import eis.iilang.Parameter;
 import jnibwapi.JNIBWAPI;
 import jnibwapi.Unit;
 
-import java.util.LinkedList;
-
-public class UnloadUnit extends StarcraftAction {
+public class UnloadUnit extends StarcraftLoadingAction {
 
   public UnloadUnit(JNIBWAPI api) {
     super(api);
-  }
-
-  @Override
-  public boolean isValid(Action action) {
-    LinkedList<Parameter> parameters = action.getParameters();
-    return parameters.size() == 1 && parameters.get(0) instanceof Numeral;
-  }
-
-  @Override
-  public boolean canExecute(Unit unit, Action action) {
-    return unit.getType().getSpaceProvided() > 0;
   }
 
   @Override
